ported tests from django_constance_cli
