+++ constance/management/commands/constance.py
@@ -8,7 +8,7 @@ from django.core.exceptions import ValidationError
 from django.core.management import BaseCommand, CommandError
 from django.utils.translation import ugettext as _
 
-from ..utils import get_constance_values, get_constance_value, set_constance_value
+from ...utils import get_constance_values, get_constance_value, set_constance_value
 
 
 class Command(BaseCommand):
